agent/requirements.txt
google-adk
google-adk[eval]
